requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.2
Flask-Migrate>=4.0.4
Flask-Login>=0.6.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Werkzeug>=2.3.7
